feat: add executable file
WIP: It would be great to run these tests as part of regular CI
in a language agnostic way.  This PR adds an executable to your path
called  `ipfs-interop` which runs the interop tests.  You should be
able to control which go/js implementations are used via the usual
IPFS_EXEC_XXX environmental variables.