refactor: make ipfs and http client peer dependencies
By depending on a certain version of ipfs and the ipfs http client
it means we can't run all the tests on other versions of ipfs/the
http client as the `require`s resolve to the wrong version.

This PR makes them peer dependencies to allow the tester/testing
environment to specify which versions of modules we want to test.