fix: throw if no requested address is available
We loop through available multiaddrs looking for certain types of
address in our tests.  If there are non of the type we are looking
for, we don't return anything which results in undescriptive
error messages from libp2p.

The change here is to throw an error with an explicit message if,
for example, no webrtc-star address is available when we expect
there to be.